+++ src/rx-collection.js
@@ -231,9 +231,11 @@ export class RxCollection {
     }
 
     /**
-     * [overwrite description]
+     * every write on the pouchdb
+     * is tunneld throught this function
      * @param {object} obj
      * @param {boolean} [overwrite=false] if true, it will overwrite existing document
+     * @return {Promise}
      */
     async _pouchPut(obj, overwrite = false) {
         obj = this._handleToPouch(obj);